Added tmux pane cycling and vimrc update
Cycle tmux panes with prefix-tab and prefix-ctrl-tab (repeatable...!).

Also added used autocmds in .vimrc so indentations are the correct size
in yml and js files (keywords should be javascript and yaml otherwise
the file types are not correctly recognized, should check the filetype
plugin for the default definitions at some later point in time.

Also g:netrw_usetab and g:netrw_wiw definitely don't work as advertised,
tried updating the netrw version but the only alpha version available
(v165) had so many bugs it was basically unusable. It did however fix
the bug that made files not renameable if the netrw listing was using
tree mode. It seemed better at detecting the current netrw dir with this
tree mode, too. It was installed under ~/.vim using a vimball and
reinstalling vim and vim-X11 did nothing because the configs were saved
in the user directory (as expected! not that I didn't realize this or
anything...!). A netrw-clean plugin correctly detected the dir and in
fact removed the updated version only cleanly.

I wonder if this g:netrw_usetab problem has to do with the X server or
maybe tmux or maybe gnome-terminal eating up the ctrl+tab key
combination. Should definitely check all shortcuts of gnome-terminal.