Added new workflows for vim and tmux
Windows (actually panes) in termux are to be selected using <prefix>q
followed by pane number. This was suggested by a disproprortionately
upvoted answer in stackoverflow where the OP was asking how to get pane
number for selecting just like you can do with a window or something
like that.

In vim I started using the leader key for personal key mappings.
Searched about the [n][v]map, [n][v]remap, [n][v]noremap similarities
and differences. It's way more intuitive than it seems as long as you
find a simple enough explanation (in stackoverflow, not in the vim's
manual...).

The first use case for this personalized shortcut was creating a
shortcut with <leader><leader> (or \\ in my case) that shrinks the
current window horizontally to about 15% of the screen size. Super
convenient for making the netrw tree file listing work like other text
editors. It's now simple to open new windows with preview (p) and a
quick resize with \\. Multiple windows can be opened with <CR> and then
quickly closed and correctly resized again.

This is related to a previous problem where g:netrw_usetab and g:netrw_wiw
were not working as described. Actually it was stated in the documentation the
reason why it wasn't working: the terminal doesn't support ctrl+tab.  More
generally, gnome doesn't support it, not just gnome-terminal. The tab key is
described ironically in a bug report as a holy key that can't be used as an
accelerator. The reasoning given was that gtk applications use tab and
its other combinations as accessibility navigators so I can't be
completely mad about it. Applications that do have these shortcuts are
simply not gtk applications.

Also included json as a filetype for 2-space indentation syntax. Writing
long semi-useless commit messages seems therapeutic. Probably it's a
good practice for basic vim editing. And simultaneously it's a training
session for touch-typing in my new setup with better posture.