Added .bashrc, .tmux.conf and .vimrc QoL improvements
Aliases:
For docker-compose up and down and for editing .rc files for
bash, vim and tmux: dup, ddown, vrc, brc, trc

tmux:
Use <prefix>q number to select specific pane, but starting with 1 so
that it is easy to reach with the left hand also binded <prefix>\ for
setting up a given window with a big main pane(for vim) and three
small ancillary bottom panes.

vim:
Started going crazy with the normal mode <leader> bindings
\c => mapped to the clear-search-highlight previously defined command
\] => set window width so that it ends at 80 characters (does it depend
on the line numbering width tho? probably, fix when it becomes annoying
or when I start editing thousands-of-lines files {hah!})
\r => reload ~/.vimrc and displays a _persistant_ status message using
redraw tricks