Went back to using 4 spaces indentation with javascript
Also realized that live reloading the .vimrc while editing a file will
destroy some options defined by filetype such as set tabstop=4 because
they have a line that overrides the filetype plugins when .vimrc is
sourced. One solution is to just do a single :e after reloading the file
or maybe later include this in the <leader>r mapping.